description field added in new tag popup + tooltip column header
Former-commit-id: fa134ffcab7fa43281ba5d6bddaa44181354b220

+++ src/modules/MainWindow/Main_Window.py
@@ -170,7 +170,8 @@ class Main_Window(QMainWindow):
             can_exit = True
 
         if can_exit:
-            shutil.rmtree(os.path.join(os.path.relpath(os.curdir), '..', '..', 'temp_project')) # temp_project removed when closing
+            if(os.path.exists(os.path.join(os.path.relpath(os.curdir), '..', '..', 'temp_project'))):
+                shutil.rmtree(os.path.join(os.path.relpath(os.curdir), '..', '..', 'temp_project')) # temp_project removed when closing
             event.accept()
         else:
             event.ignore()
